tests/web_platform_tests/graph/governance-constraints.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Graph Governance — Constraint Enforcement — WPT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

// ==========================================================================
// Spec 05 — Graph Governance: Constraint Enforcement for Shared Linked Data Graphs
// https://github.com/HexaField/w3c-living-web-proposals/blob/main/drafts/05_graph-governance.md
// ==========================================================================

// §9 Governance Engine API

promise_test(async t => {
  const graph = await navigator.graph.create('GovAPI');
  const shared = await graph.share();

  assert_true(typeof shared.canAddTriple === 'function', 'canAddTriple MUST exist');
  assert_true(typeof shared.constraintsFor === 'function', 'constraintsFor MUST exist');
  assert_true(typeof shared.myCapabilities === 'function', 'myCapabilities MUST exist');

  await navigator.graph.remove(graph.uuid);
}, '§9 SharedGraph MUST expose canAddTriple(), constraintsFor(), myCapabilities()');

// §9.1 canAddTriple()

promise_test(async t => {
  const graph = await navigator.graph.create('CanAdd');
  const shared = await graph.share();

  // With no constraints, all triples should be allowed
  const triple = new SemanticTriple('ex:a', 'ex:b', 'app://body');
  const result = await shared.canAddTriple(triple);
  assert_true(result.allowed === true || result === true,
    'MUST allow when no constraints exist');

  await navigator.graph.remove(graph.uuid);
}, '§9.1 canAddTriple() MUST return allowed:true when no constraints exist');

promise_test(async t => {
  const graph = await navigator.graph.create('CanAddResult');
  const shared = await graph.share();

  const triple = new SemanticTriple('ex:a', 'ex:b', 'app://body');
  const result = await shared.canAddTriple(triple);

  // ValidationResult shape check
  if (typeof result === 'object' && result !== null) {
    assert_true('allowed' in result, 'MUST have allowed field');
    if (!result.allowed) {
      assert_true('module' in result, 'rejected results MUST have module');
      assert_true('reason' in result, 'rejected results MUST have reason');
    }
  }

  await navigator.graph.remove(graph.uuid);
}, '§9.1 canAddTriple() MUST return ValidationResult with allowed, module, reason');

// §9.2 constraintsFor()

promise_test(async t => {
  const graph = await navigator.graph.create('Constraints');
  const shared = await graph.share();

  const constraints = await shared.constraintsFor();
  assert_true(constraints !== null && constraints !== undefined,
    'MUST return a value');
  // May be empty array if no constraints
  if (Array.isArray(constraints)) {
    assert_true(true, 'returns array of constraints');
  }

  await navigator.graph.remove(graph.uuid);
}, '§9.2 constraintsFor() MUST return constraints for the given scope');

promise_test(async t => {
  const graph = await navigator.graph.create('ConstraintScope');
  const shared = await graph.share();

  // constraintsFor with specific entity
  const constraints = await shared.constraintsFor('some-entity');
  assert_true(constraints !== null && constraints !== undefined);

  await navigator.graph.remove(graph.uuid);
}, '§9.2 constraintsFor(entityAddress) MUST accept an entity address parameter');

// §9.3 myCapabilities()

promise_test(async t => {
  const graph = await navigator.graph.create('MyCaps');
  const shared = await graph.share();

  const caps = await shared.myCapabilities();
  assert_true(caps !== null && caps !== undefined, 'MUST return capabilities');

  // As graph creator (root authority), should have capabilities
  if (Array.isArray(caps)) {
    // Root authority may have implicit capabilities
    assert_true(true, 'returns array of capabilities');
  }

  await navigator.graph.remove(graph.uuid);
}, '§9.3 myCapabilities() MUST return capabilities for current user');

// §4.3 Capability constraints

promise_test(async t => {
  const graph = await navigator.graph.create('CapConstraint');
  const shared = await graph.share();

  // Add a capability constraint as triples
  await shared.addTriple(new SemanticTriple(
    shared.uri, 'urn:constraint:cap1', 'governance://has_constraint'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:cap1', 'governance://constraint', 'governance://entry_type'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:cap1', 'capability', 'governance://constraint_kind'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:cap1', 'required', 'governance://capability_enforcement'));

  // Root authority should still be allowed
  const triple = new SemanticTriple(shared.uri, 'test', 'app://body');
  const result = await shared.canAddTriple(triple);
  // Root authority has implicit capability
  assert_true(result.allowed === true || result === true,
    'Root authority MUST be allowed even with capability constraint');

  await navigator.graph.remove(graph.uuid);
}, '§4.3 Root authority MUST have implicit capability over all predicates');

// §4.5 Temporal constraints

promise_test(async t => {
  const graph = await navigator.graph.create('Temporal');
  const shared = await graph.share();

  // Add a temporal constraint
  await shared.addTriple(new SemanticTriple(
    shared.uri, 'urn:constraint:rate1', 'governance://has_constraint'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:rate1', 'governance://constraint', 'governance://entry_type'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:rate1', 'temporal', 'governance://constraint_kind'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:rate1', '30', 'governance://temporal_min_interval_seconds'));

  const constraints = await shared.constraintsFor();
  // Should include the temporal constraint
  assert_true(constraints !== null);

  await navigator.graph.remove(graph.uuid);
}, '§4.5 Temporal constraints MUST be queryable via constraintsFor()');

// §4.6 Content constraints

promise_test(async t => {
  const graph = await navigator.graph.create('Content');
  const shared = await graph.share();

  // Add a content constraint
  await shared.addTriple(new SemanticTriple(
    shared.uri, 'urn:constraint:content1', 'governance://has_constraint'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:content1', 'governance://constraint', 'governance://entry_type'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:content1', 'content', 'governance://constraint_kind'));
  await shared.addTriple(new SemanticTriple(
    'urn:constraint:content1', '200', 'governance://content_max_length'));

  const constraints = await shared.constraintsFor();
  assert_true(constraints !== null);

  await navigator.graph.remove(graph.uuid);
}, '§4.6 Content constraints MUST be queryable via constraintsFor()');
</script>
